requirements.txt
faster-whisper
pygame
playsound
SpeechRecognition
edge_tts
google-generativeai